refactor(entity): remove redundant explicit double conversions in fromJson methods
Simplify numeric field parsing by removing unnecessary `.toDouble()` casts on JSON values
that already expect doubles, relying on the `??` operator to provide the correct type.
This avoids unnecessary explicit conversions and makes the code cleaner.

+++ lib/entity/creature_model_info_entity.dart
@@ -16,8 +16,8 @@ class CreatureModelInfoEntity {
   factory CreatureModelInfoEntity.fromJson(Map<String, dynamic> json) {
     return CreatureModelInfoEntity(
       displayId: json['DisplayID'] ?? 0,
-      boundingRadius: (json['BoundingRadius'] ?? 0).toDouble(),
-      combatReach: (json['CombatReach'] ?? 0).toDouble(),
+      boundingRadius: (json['BoundingRadius'] ?? 0.0),
+      combatReach: (json['CombatReach'] ?? 0.0),
       gender: json['Gender'] ?? 0,
       displayIdOtherGender: json['DisplayID_Other_Gender'] ?? 0,
     );
